requirements.txt
absl-py==0.9.0
appdirs==1.4.4
argcomplete==1.12.0
asgiref==3.2.10
audioread==2.1.8
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
densetorch==0.0.2
Django==3.1
docopt==0.6.2
dominate==2.5.1
dtw==1.4.0
dtw-c==0.7.0
fastdtw==0.3.4
filelock==3.0.12
future==0.18.2
future-fstrings==1.2.0
gdown==3.12.2
google-auth==1.20.1
google-auth-oauthlib==0.4.1
GPUtil==1.4.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imgaug==0.2.6
importlib-metadata==1.7.0
joblib==0.16.0
jsonpatch==1.26
jsonpointer==2.0
kaldiio==2.15.1
kiwisolver==1.2.0
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.2.2
matplotlib==3.3.0
mlpg-c==0.0.7.dev1
networkx==2.4
numba==0.48.0
numpy==1.19.1
oauthlib==3.1.0
opencv-python-headless==4.4.0.44
packaging==20.4
-e git+https://github.com/kan-bayashi/ParallelWaveGAN.git@69d22bd780708b8603590b3f6c37a91231849bf0#egg=parallel_wavegan
Pillow==7.2.0
pooch==1.1.1
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
pysptk==0.1.18
pytest-depends==1.0.1
pytest-mock==3.3.1
pytest-repeat==0.8.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyworld==0.2.10
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
Resemblyzer==0.1.1.dev0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
SoundFile==0.10.3.post1
sprocket-vc==0.18.3
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
threadpoolctl==2.1.0
tifffile==2020.8.13
torch==1.6.0+cu101
torchaudio==0.6.0
torchfile==0.1.0
torchvision==0.7.0+cu101
tornado==6.0.4
tqdm==4.48.2
typing==3.7.4.3
umap==0.1.1
urllib3==1.25.10
visdom==0.1.8.9
wavenet-vocoder==0.1.1
webrtcvad==2.0.10
websocket-client==0.57.0
Werkzeug==1.0.1
xmltodict==0.12.0
yq==2.10.1
zipp==3.1.0
